fix all existing tests
