+++ tests/cmds/conftest.py
@@ -117,3 +117,9 @@ def get_generator_for_get_all(mocker, mock_return_items):
         yield Py42Response(response)
 
     return gen
+
+
+def get_mark_for_search_and_send_to(command_group):
+    search_cmd = [command_group, "search"]
+    send_to_cmd = [command_group, "send-to", "0.0.0.0"]
+    return pytest.mark.parametrize("command", (search_cmd, send_to_cmd))
