+++ tests/cmds/test_auditlogs.py
@@ -6,6 +6,7 @@ from logging import Logger
 import pytest
 from py42.response import Py42Response
 from requests import Response
+from tests.cmds.conftest import get_mark_for_search_and_send_to
 
 from code42cli.cmds.auditlogs import _parse_audit_log_timestamp_string_to_timestamp
 from code42cli.cmds.search.cursor_store import AuditLogCursorStore
@@ -60,6 +61,7 @@ TEST_EVENTS_WITH_DIFFERENT_TIMESTAMPS = [
 TEST_CHECKPOINT_EVENT_HASHLIST = [
     hash_event(event) for event in TEST_EVENTS_WITH_SAME_TIMESTAMP
 ]
+search_and_send_to_test = get_mark_for_search_and_send_to("audit-logs")
 
 
 @pytest.fixture
@@ -133,17 +135,20 @@ def test_audit_log_response_with_only_same_timestamps(mocker):
     return response_gen()
 
 
-def test_search_audit_logs_json_format(runner, cli_state, date_str):
-    runner.invoke(cli, ["audit-logs", "search", "-b", date_str], obj=cli_state)
+@search_and_send_to_test
+def test_search_and_send_to_handles_json_format(runner, cli_state, date_str, command):
+    runner.invoke(cli, [*command, "-b", date_str], obj=cli_state)
     assert cli_state.sdk.auditlogs.get_all.call_count == 1
 
 
-def test_search_audit_logs_with_filter_parameters(runner, cli_state, date_str):
+@search_and_send_to_test
+def test_search_and_send_to_handles_filter_parameters(
+    runner, cli_state, date_str, command
+):
     runner.invoke(
         cli,
         [
-            "audit-logs",
-            "search",
+            *command,
             "--actor-username",
             "test@test.com",
             "--actor-username",
@@ -166,13 +171,15 @@ def test_search_audit_logs_with_filter_parameters(runner, cli_state, date_str):
     )
 
 
-def test_search_audit_logs_with_all_filter_parameters(runner, cli_state, date_str):
+@search_and_send_to_test
+def test_search_and_send_to_handles_all_filter_parameters(
+    runner, cli_state, date_str, command
+):
     end_time = datetime.utcnow().strftime("%Y-%m-%d %H:%M:%S")
     runner.invoke(
         cli,
         [
-            "audit-logs",
-            "search",
+            *command,
             "--actor-username",
             "test@test.com",
             "--actor-username",
@@ -244,42 +251,17 @@ def test_send_to_emits_events_in_chronological_order(
     )
 
 
-def test_search_with_checkpoint_saves_expected_cursor_timestamp(
-    cli_state, runner, test_audit_log_response, audit_log_cursor_with_checkpoint
-):
-    cli_state.sdk.auditlogs.get_all.return_value = test_audit_log_response
-    runner.invoke(
-        cli,
-        ["audit-logs", "search", "--begin", "1d", "--use-checkpoint", "test"],
-        obj=cli_state,
-    )
-    assert audit_log_cursor_with_checkpoint.replace.call_count == 4
-    assert audit_log_cursor_with_checkpoint.replace.call_args_list[3][0] == (
-        "test",
-        CURSOR_TIMESTAMP,
-    )
-
-
-def test_send_to_with_checkpoint_saves_expected_cursor_timestamp(
+@search_and_send_to_test
+def test_search_and_send_to_with_checkpoint_saves_expected_cursor_timestamp(
     cli_state,
     runner,
     test_audit_log_response,
     audit_log_cursor_with_checkpoint,
-    send_to_logger,
+    command,
 ):
     cli_state.sdk.auditlogs.get_all.return_value = test_audit_log_response
     runner.invoke(
-        cli,
-        [
-            "audit-logs",
-            "send-to",
-            "localhost",
-            "--begin",
-            "1d",
-            "--use-checkpoint",
-            "test",
-        ],
-        obj=cli_state,
+        cli, [*command, "--begin", "1d", "--use-checkpoint", "test"], obj=cli_state,
     )
     assert audit_log_cursor_with_checkpoint.replace.call_count == 4
     assert audit_log_cursor_with_checkpoint.replace.call_args_list[3][0] == (
@@ -288,34 +270,16 @@ def test_send_to_with_checkpoint_saves_expected_cursor_timestamp(
     )
 
 
-def test_search_with_existing_checkpoint_replaces_begin_arg_if_passed(
-    cli_state, runner, test_audit_log_response, audit_log_cursor_with_checkpoint
+@search_and_send_to_test
+def test_search_and_send_to_with_existing_checkpoint_replaces_begin_arg_if_passed(
+    cli_state,
+    runner,
+    test_audit_log_response,
+    audit_log_cursor_with_checkpoint,
+    command,
 ):
     runner.invoke(
-        cli,
-        ["audit-logs", "search", "--begin", "1d", "--use-checkpoint", "test"],
-        obj=cli_state,
-    )
-    assert (
-        cli_state.sdk.auditlogs.get_all.call_args[1]["begin_time"] == CURSOR_TIMESTAMP
-    )
-
-
-def test_send_to_with_existing_checkpoint_replaces_begin_arg_if_passed(
-    cli_state, runner, test_audit_log_response, audit_log_cursor_with_checkpoint
-):
-    runner.invoke(
-        cli,
-        [
-            "audit-logs",
-            "send-to",
-            "localhost",
-            "--begin",
-            "1d",
-            "--use-checkpoint",
-            "test",
-        ],
-        obj=cli_state,
+        cli, [*command, "--begin", "1d", "--use-checkpoint", "test"], obj=cli_state,
     )
     assert (
         cli_state.sdk.auditlogs.get_all.call_args[1]["begin_time"] == CURSOR_TIMESTAMP
@@ -338,74 +302,19 @@ def test_search_with_existing_checkpoint_events_skips_duplicate_events(
     assert "43@code42.com" in result.stdout
 
 
-def test_send_to_with_existing_checkpoint_events_skips_duplicate_events(
-    cli_state,
-    runner,
-    test_audit_log_response,
-    audit_log_cursor_with_checkpoint_and_events,
-    send_to_logger,
-):
-    cli_state.sdk.auditlogs.get_all.return_value = test_audit_log_response
-    runner.invoke(
-        cli,
-        [
-            "audit-logs",
-            "send-to",
-            "localhost",
-            "--begin",
-            "1d",
-            "--use-checkpoint",
-            "test",
-        ],
-        obj=cli_state,
-    )
-    assert send_to_logger.info.call_count == 3
-    assert send_to_logger.info.call_args_list[0][0][0]["actorName"] != "42@code42.com"
-
-
-def test_search_without_existing_checkpoint_writes_both_event_hashes_with_same_timestamp(
+@search_and_send_to_test
+def test_search_and_send_to_without_existing_checkpoint_writes_both_event_hashes_with_same_timestamp(
     cli_state,
     runner,
     test_audit_log_response_with_only_same_timestamps,
     audit_log_cursor_with_checkpoint,
+    command,
 ):
     cli_state.sdk.auditlogs.get_all.return_value = (
         test_audit_log_response_with_only_same_timestamps
     )
     runner.invoke(
-        cli,
-        ["audit-logs", "search", "--begin", "1d", "--use-checkpoint", "test"],
-        obj=cli_state,
-    )
-    assert audit_log_cursor_with_checkpoint.replace_events.call_count == 2
-    assert audit_log_cursor_with_checkpoint.replace_events.call_args_list[1][0][1] == [
-        hash_event(TEST_EVENTS_WITH_SAME_TIMESTAMP[0]),
-        hash_event(TEST_EVENTS_WITH_SAME_TIMESTAMP[1]),
-    ]
-
-
-def test_send_to_without_existing_checkpoint_writes_both_event_hashes_with_same_timestamp(
-    cli_state,
-    runner,
-    test_audit_log_response_with_only_same_timestamps,
-    audit_log_cursor_with_checkpoint,
-    send_to_logger,
-):
-    cli_state.sdk.auditlogs.get_all.return_value = (
-        test_audit_log_response_with_only_same_timestamps
-    )
-    runner.invoke(
-        cli,
-        [
-            "audit-logs",
-            "send-to",
-            "localhost",
-            "--begin",
-            "1d",
-            "--use-checkpoint",
-            "test",
-        ],
-        obj=cli_state,
+        cli, [*command, "--begin", "1d", "--use-checkpoint", "test"], obj=cli_state,
     )
     assert audit_log_cursor_with_checkpoint.replace_events.call_count == 2
     assert audit_log_cursor_with_checkpoint.replace_events.call_args_list[1][0][1] == [
